Add details for making native app multi-tenant
By default, Azure creates new native apps with visibility only to the tenant
who owns the app registration. Unlike the web app, Azure does not provide a
straightforward toggle for this.

For this sample to work in its entirety (i.e. against another tenant), an
additional step is required to configure the native app for multitenancy.